LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/revert-val-008-expected.txt


FAIL The revert keyword works in the final frame of a web animation assert_between_exclusive: expected a number greater than 21 and less than 42 but got 15
